Finish up compiletest.py
Now tries to build on my machine with gcc 4.1-4.8 (can be easily extended
if needed) which is the current range we want to support.

Other platforms only implemented tentatively, may be broken.

Currently requires a tracelib-package-install which means this won't yet
work with the ci-job unfortunately. Will have to think about how we can
fix that (maybe always pass ARCH_LIB_SUFFIX or so)