Consider tracepoint activeness before visiting it
This fixes getting output for a tracepoint even if it is deactivated, but
without messages or variables. The reason for this is that the macro
expansion for the streaming case yields a visitor-object outside the
if(advanceVisit) check. The streaming is completely inside the if, so it
does not happen if the tracepoint is not active. However the visitor is
destructed eventually and hence it'll flush its tracepoint unconditionally.

Fix that by considering the active flag of the tracepoint for flushing.