Replace TracepointVisitorProxy with an explicit end marker
This is necessary as different compiler generate differing code for the
streaming proxy stuff. In particular MSVC generates code that calls the
copy constructor of the proxy and hence generates deletion of one more
instance than we want/expect. That means with MSVC one got each trace
using the streaming syntax listed multiple times in the output (visited
multiple times).

The only viable fix for this is to require an explicit end marker that
'flushes' the visitor and visits the tracepoint.