Fix multiple loads of the config files triggerd by the monitor
The monitor decided to emit a change-notification twice when initially
starting. The reason for that is that the initial values for fileExists
and lastModificationTime where hardcoded, which means if the config
file existed already when the monitor started it noticed the existences and
notified about that and then also noticed the difference in modification
time because the start value was 0.

Fix that by initializing both variables before entering the loop once based
on the 'current' situation on disk.

Avoid code-duplication by using a simple struct for holding and retrieving
the data.