de-inline size/operator[] of variablesnapshot
The thesis is that inline functions let the compiler make assumptions
about the binary layout of the objects. Since the compiler that builds
tracelib is a different one than the one using tracelib this is not a good
idea at least with msvc which regularly breaks binary compatibility.

Uninlining these functions fixes a crash I can reproduce when running
compiletest.cpp built with msvc8 against tracelib built with msvc9. That
matches the above outlined theory.