Consider signed/unsigned char as integral type not character
This fixes differences between behavior for int8_t in msvc and gcc. The
latter considered it a char producing no character (since initialized to
0) for it while msvc uses one of the integral overloads instead and
produces a number.

This change makes gcc consider int8_t a numerical value as well since
its typedef'ed to signed char (and uint8_t respectively). This should be
ok since most people will use plain 'char' as type for characters and not
explicitly make it signed or unsigned.

Adapt the expected results.