Fix copy-constructor for VariableValue
Not copying the isSigned member meant undefined behaviour when copying
VariableValue's. Such a copy is created when using convertVariable which
means almost always when dealing with that type.

Interestingly many compilers/compiler combinations initialized the variable
to true anyway which made it hard to find the problem. Only the combination
of building tracelib with msvc10 (and as debug build) and compiling squish
against it with msvc10 as well (a msvc9 squish build worked) exhibited the
issue by producing big positive values for signed integers with the value -1.