Fix handling of messages with percent characters
QString::arg is unable to handle strings with percent in them due to the
way its applied. Hence replace the usage of that with simple string
concatenation. This could be a bit slower, but since its only a temporary
fix should be safe enough.

Ultimately this stuff should use prepared statements to let the DB/driver
do escaping etc.