Fixed occasional crash in clients built with MSVC11
For some obscure program, client programs built with Visual Studio 2012 would
occasionally(!) crash when using tracelib. It turned out that a sample program
as simple as

  #include <tracelib/tracelib.h>

  int main()
  {
      froglogic::tracelib::VariableSnapshot *snapshot
          = new froglogic::tracelib::VariableSnapshot;
      return 0;
  }

*sometimes* crashes - but only with MSVC11. Neither MSVC9 nor MSVC12 exhibited
this problem.

It apperas that this is somehow related to how the std::vector object
internally manages memory in Visual Studio 2012: if we make sure to allocate it
on the heap, it appears to work alright. I wrapped the pointer into a
std::auto_ptr just so that I cannot forget to delete it.

This patch seems to avoid the crash without uglifying the code too much.