Try to fix race between post-compiler-write and removal of executable
Turns out that at least on reinecke there is a race that happens quite
consistently between the compiler-started processes to write extra struff
in the executable, that executable starting up and the removal of the
executable before recompiling.

Ultimately we really need to create separate builddirs for this but as a
quick fix for now just sleep a bit until we can write the file.