Lazily create the stringstream
Profiling showed that the constructor of TracePointVisitor was the major
pain point when using fTrace without actually producing a trace file/output.
The constructor itself is almost empty, what took the time is constructing
the stringstream.

Since the visitor is created even when the tracepoint is not visited, its
construction has to be as cheap as possible so create the stream lazily.