Fix printing of QMap<QString,QVariant>
QVariant does not support toString for the Map type, so lets print this
ourselves with a small helper that also supports printing recursively.

We can easily extend this in case other types are not printed with useful
values.

Revert the QMap<QString,QString> variant to use the QVariant overload again
now that that one produces proper out.

Slightly changed the output to not use quotes around the value since that
may now be a map again.