Properly 'escape' CDATA content
The end token of a CDATA section (]]>) is not allowed inside the section
and cannot be escaped either. Instead one has to split the content so that
one part of the end token is part of one CDATA section and the rest is in
the next one. Do this for all places that produce CDATA even though some
of them are extremely unlikely to produce such end tokens.

This also helps the xml2trace tool since the invalid xml made it abort
the conversion.