Fix production of process shutdown event
The logging of the shutdown event is very late in the process shutdown
and hence the std::string objects may have been deleted already on some
platforms. This led to empty-string being searched for which is found the
first time at position 0 and hence the first 3 characters of the process
name where cut off.

Since we don't really need a C++ string object for these constants, lets
just use a C string to make sure the deletion happens later.