Properly initialize modification time for unix polling observer
This one has the same issue that I fixed in the windows polling code
a couple of months ago. The modification_time is not initialized when the
observer is created and hence upon the first time the timer fires the code
thinks the file has 'just appeared'. Thats of course wrong if the file
already existed and with the current behavior of tracelib truncating the
file upon each configuration reload its a real pain as the first 5 seconds
will be lost.

This was found on Mac but may also happen on Linux systems that have no
inotify.