Default to enabling all trace group's in the GUI
I think this is a saner default as one can easily miss the filtering option
when using this tool the first time and having a trace that only has entries
with a corresponding group associated. The list is just empty and makes the
user wonder wether there was some error somewhere.