Switch tracelib.xml to specify the basename of the executable
Instead of having to add/remove the executable suffix when moving the
config file from one platform to another lets simply enforce that the
specified process name is just the base name of the process executable
without the suffix.

I thought about providing some backwards compatibility options but that
gets hard to do when .exe is specified and using the config on *nix and
since we're the only user of this project right now I'll just change
the format.