Rename & Relicense source code under LGPL v3
This renames the framework to 'tracetool', which is a little broader
than 'tracelib' and which doesn't seem to be prone to trademark issues.
The patch also relicenses all parts of the tracetool framework to be
subject to the Lesser General Public License version 3.

Prior versions of the framework were purely meant for internal usage,
and froglogic reserved all rights to the software; the lack of an
explicit license statement did not imply that the software was in the
Public Domain.