Rename & Relicense further source code under LGPL v3
I missed a couple of files in my previous commit which relicensed the source
code; this commit continues the work, relicensing more parts of the tracetool
framework to be subject to the Lesser General Public License version 3.

Prior versions of the framework were purely meant for internal usage, and
froglogic reserved all rights to the software; the lack of an explicit license
statement did not imply that the software was in the Public Domain.