Replace GetOpt with QCommandLineParser in the gui application
The qt5 port of GetOpt seems to be a little buggy, the optional argument
for the tracefile gets filled with argv[0] when calling tracegui without
arguments. Instead of trying to debug GetOpt I think its nicer to just
switch to a Qt-provided parser. That also has the benefit of a slightly
nicer API, taking over even more work then GetOpt (handling of --help
and --version) and not having to figure out what to pass as arguments list
to it.