Port the commandline tools away from getopt
GetOpt parsing with Qt5 has a bug where the first positional argument
gets assigned the program executable path instead of the actual argument.
So switch the commandline tools to QCommandLineParser which also reduces
the amount of code needed quite a bit.

Wherever the tool previously would complain about missing arguments I
opted for directly including the usage-output instead of telling the user
to run again with --help.

Also added the --version switch already even though the output will not
include a version because the code does not call
QCoreApplication.setApplicationVersion.

Last but not least I opted for 'fixing' the illogical behavior of xml2trace
so that it works not exactly like its companion trace2xml but rather
mirrors its behavior. That is instead of specifying the trace db file via
a --output option the command now expects the db as positional argument.
Instead the xml input file that is optional is specified now via --input
(or still possible: use stdin). That makes more sense since the -o argument
was never optional as outputting a db to stdout makes little to no sense.