Fix build instructions to use the proper cmake syntax
The Qt5 cmake files are placed in a special subdirectory that CMake will
search for automatically as long as the Qt5 directory is part of its
list of 'prefixes'. By default those are /usr/ and /usr/local as well as
the install prefix (IIRC) on *nix, but one can specify additional paths
on the commandline.

The benefit of this is that one has to remember only 1 variable and
it'll work for all projects providing cmake config modules instead of
having to remember special variables like Qt5_Dir.