Require CMake 3.0 to gain automatic linking of qtmain.lib
CMake 2.8.11 introduce automatic linking of the qtmain.lib file and 3.0 was
released 2 years ago. Since we'll likely require Qt5.5 or Qt5.6 as minimum
in this branch anyway lets update to a more recent CMake version to silence
the policy-warning CMP0020 on windows and avoid having to run around all
the CMakeListst.txt to manually fix the linking of the qtmain.lib so that
the WinMain function is provided in the executables.