Add a travis-ci configuration for building tracetool
This builds tracetool on Ubuntu Trusty 14.04 since older versions do not
allow using Qt5 on them apparently. I chose ninja as build tool to
automatically make use of multiple cores and speed up the build. Most
time is still spent in installing the Qt dependencies.

The qttools* packages are needed for linguist which supports translating
of the UI of tracelib.