Give the failing pcretest a bit more memory
It seems the default size is a bit too small when the code is compiled with
clang to run the test successfully. Just increasing the limit of the stack
is enough for the test to succeeds so do that since the test itself claims
it requires a relatively large stack.