Disable the double-click handler for macOS
For a currently unknown reason showing the message box triggers a qFatal
on macOS saying that one needs to create a QApplication and not
QCoreApplication. However the application does create a QApplication
object already, so no idea why Qt thinks that.

Since running into a crash every time I'm accidentally double-clicking
somewhere is pretty annoying, lets disable this for now on macOS (I think
showing the backtrace if there's one in a message box is questionable
anyway).