Improve Backtrace Information Dialog
when viewing the backtraces one might want to copy the text, but the text wasn't
selectable, so i added the proper text interaction flag.
Since the QMessageBox has some weird resize restrictions and wouldn't fully
expand to show the entire backtrace i had to resort to this weird work around.