Support packaging just the hook library
We may need this to be able to build the parts loaded into applications
on systems that run rather old versions of their operating systems. So
it would be useful to be able to create the packages related to that
without the need to have a Qt installation on the system.