Remove warning about type incompatibility for string formatting
This removes a warning that appears in users of the project which is quite
annoying. The best option would be to use '%td' for ptrdiff_t, but that is
not supported by all versions of MSVC. So use that or the msvc-specific
format specifier depending on the compiler being used.