Ensure to add rpath's to the install location
At least on macOS Qt5 comes with the framework's install name set to an
rpath-relative path. Hence the binaries linked against Qt require to have
an rpath add to them. This option enables cmake to do exactly that, it'll
add all 'external' non-system link directories to the rpath. In the case
of tracetool that's just the Qt directory.

I'm not sure how this affects Linux or generated packages yet, but I don't
expect any extra harm - compared to having no rpath at all.