Allow to disable CMAKE_INSTALL_RPATH_USE_LINK_PATH for package builds
The setting would leak paths from the system where the package is built
into the packages as cmake will do an install step before creating the
packages and the deployment tools will also adjust the rpath as necessary.
So at least on macOS the executables end up with the
@executable_path-setting for redistribution as well as the original Qt
installation path, which is not nice.