Make sure that windeployqt is also invoked during packaging
CPack uses component-based packaging, so the windeployqt invovation needs
to be assigned to a component.

In addition it has to resolve the CMAKE_INSTALL_PREFIX variable during its
own invocation, not during the cmake invocation configuring the build
directory. CPack ends up specifying a different install directory when
running the install command, so the value of the variable changes when
cpack actually installs the component.

This fixes inclusion of the Qt dll's inside the built packages.