Bundle Qt5 for Linux
On Linux Qt does not provide a deployment tool so we have to do this
manually, though the basic idea (copy libs, plugins, set rpath and create
qt.conf) is the same as on other platforms.

This attempts to use BundleUtilities to do the heavy lifting of figuring
out the dependencies of Qt. The plugins need to be copied manually,
BundleUtilities does not support doing that. Same goes for creating a
qt.conf.

Unfortunately the resulting installation is not usable because:
* libraries are copied into the bin/ folder where the exe is
* runpath is stripped from the Qt libraries leaving the libraries broken
  as they depend on each other

I'm committing this nonetheless since I'd prefer this if we can fix
BundleUtilities than implementing the copying manually with a list of
fixed libraries.

In order to use this one needs to pass CMAKE_INSTALL_RPATH=\$ORIGIN/../lib
and BUNDLE_QT=ON to cmake to enable the bundling of qt and ensuring the
binaries have an rpath allowing them to find qt inside the lib/ folder.