Manually copy Qt libraries to the install location when packaging
Since the BundleUtilities invocation won't work (see previous commit) lets
copy the files manually to the right places. This is less magic at the
expense of more code and no automatic handling of changing dependencies.
For the plugins I've used the targets, but for the libraries that wasn't
doable since the target's only refer to the actual files but the binaries
are being linked to the so-versioned name (i.e. libQtCore.so.5 instead of
libQtCore.so.5.9.0). So copying just the actual files was insufficient and
some of the libraries had no targets anyway.