Replace python code for cppcheck with a target
This reduces the dependencies of the project, no python needed just to run
cppcheck. It'll also enable usage of information from cmake to supply
better defines/includes to cppcheck if we want to.

I've dropped the list of defines/undefines since they did not seem to be of
any use, i.e. the issues found by cppcheck did not seem to be dependent on
these.