Fix building by moving the LineEdit constructor into a separate file
While the class has little own logic I think its necessary to keep this
in a class for the ItemEditor in the table. However for automoc to work
properly its necessary that there's a cpp file and not everything inline
in the header. Otherwise the moc'ing does not happen or the sources are
not being compiled into the target.