Support bundling of Qt library on macOS as well
Use the same BUNDLE_QT flag as for linux but run macdeployqt instead of
manually copying around the framework files. Less code, but needs to be
applied to each executable.

I decided to duplicate this instead of creating a shared function as we
have no good place - except dumping it in a random place in the toplevel
cmake file - and its not that much logic anyway.

This is intended to be used with cpack to create packages, but can also
be used separately - though usually the install-rpath is sufficient for
a local installation.