Enable running CI and package builds on Linux/macOS
This extends the build configuration for travis to enable package builds
on Linux and macOS. The actual build commands are moved to shell scripts
as there's a bit more logic involved and the inline conditions are
getting hard to read.

The largest change is using a docker image to build tracetool on Linux
such that a Qt installation can be used that's fit for redistribution.

The deployment happens only for pushes to tags, which end up defining a
release and only for the clang builds (this is arbitrary chosen on my
preference for clang over gcc).

The docker image uses Qt5.9.6 which is the latest LTS release available.

The cmake invocation ensures that Qt is being bundled and that on Linux
the RPATH is adjusted to find that qt installation (on mac the
macdeployqt tool does this).

The minimum macOS version supported by the packages is 10.10, which is a
requirement of the Qt version.

For builds of tags the build will run in Release mode to ensure maximum
performance of the resulting binaries.

Tests are always being performed, ensuring that released packages did
run the automatic tests successfully.

The authentication token is bound to my github account, but can be
easily replaced whenever someone else wants to take over the
maintenance.