Fix building tracelib's hook library with gcc using visibility flags
Most modern projects use gcc's visibility support to reduce the number
of exported symbols. This is done by changing the visibility default to
not-visible and then individual functions/classes that should be visible
are marked to be visible. The export macro for the hooklib did not support
this yet, leaving all symbols invisible when enabling this feature on
Linux/Unix.