Guard the ucontext include with __sun
The api is solaris-specific so it shouldn't be used on other systems. So
far this is usually not a problem since 'other systems' generally have
the bfd-stuff and use that code-path. However if the bfd-headers are not
found the code attempted to use ucontext.h which won't work on Linux for
example.