Update appveyor configuration to do ci and package builds
A couple of changes for the windows build setup are being done here to
allow running CI builds (including tests) and also generate packages and
upload them to github releases.

The actual cmake/ninja commands are being moved to cmd files since quite
a bit of logic has been added to distinguish between release and
non-release builds as well as the two different architectures to build.

The version is being set to just the build number to avoid confusions
with the actual tracetool release version.

In order to speed up builds a bit the ninja and doxygen install's are
being cached.

Deployment uses a github token bound to my account, this can be changed
at any time allowing somebody else to take over the deployment. The
token has to be manually encrypted using appveyor's settings as github
will revoke the token if its found in the repository.

The distinction between package and non-package builds is done on
APPVEYOR_REPO_TAG which is only being set when a build is triggered by
pushing to a tag. This tag will then also be used to upload the packages
to on github, making it a fully release.

The build.cmd file is mostly just setup-boilerplate, doing a few
decisions based on the architecture to build and whether package builds
should be done or not. The build is based on Qt5.9.5, since Appveyor has
no Qt5.9.6 yet.

The package builds are running tests to ensure proper working packages
and they build in release mode to ensure maximum performance.

The test cmd file uses the fact that the environment setup by build.cmd
can just be reused to ensure the test binaries can find qt. Putting
these exact commands into the yaml-file hasn't worked, hence such a
trivial cmd file was added.