Fix position of export macro for tracepoint.h
This hasn't been a problem for tracetool so far since the export macro
on Linux was empty and it appears msvc does not complain about a
dllexport/dllimport at this position. It also seems to be ignored by newer
versions of gcc - at least I didn't get a compile error for users of this
header in my Ubuntu.

However older gcc systems (gcc 4.1 in particular) barfed about the
export macro (and hence visibility-attribute) in the line with the enum.
That's also quite unusual, normally only classes or functions are marked
for export.

In this case I suspect the export macro is of no use anyway, since the
struct is only used as 'namespace' for the enum value and has no functions.
So its a 'header only' thing and thus no export macro needed. Since the
macro does not seem to hurt though, lets keep it in case I did overlook
something.