Update access token for github for travis/appveyor
The previous token was not usable for the appveyor setup since a new
dedicated account was created for connecting the tracetool repository
under the froglogic organization with appveyor. Since the encryption uses
the account information for encrypting the change of account causes the
token to be useless.

I'm also updating the token used for travis since I'd like to avoid both
services using two different tokens.