Report xml parsing errors from xml2trace to the user
Invalid xml files may be generated by users of the trace library since
the generated messages are not sanitized and hence binary data or
non-allowed characters can spill into the xml. In addition crashes etc.
may lead to cut-off files.

The extra handling for premature-end errors is necessary since xml2trace
feeds the data in chunks, so the stream reader will constantly generate
that error when the current chunk has been read in full. I opted for
handling that internally instead of generating a dedicated error so the
calling side can decide if the premature-end is an actual error. That turned
out to be a lot of extra code (since the root-element has to be closed and
parsed again when this error occurs and the input file is at the end) for
a use-case that can be easily detected by looking at the end of the origina
log file.